bdev/iscsi: silence clang enum conversion warning
This assignment is converting from the libiscsi sense key enumerated
type to the SPDK one, and clang issues a warning for this case.  We know
that these enums are equivalent, so just add a cast to avoid the
warning.

+++ lib/bdev/iscsi/bdev_iscsi.c
@@ -214,7 +214,7 @@ bdev_iscsi_rw_cb(struct iscsi_context *context, int status, void *_task, void *_
 	struct bdev_iscsi_io *iscsi_io = _iscsi_io;
 
 	iscsi_io->scsi_status = task->status;
-	iscsi_io->sk = task->sense.key;
+	iscsi_io->sk = (uint8_t)task->sense.key;
 	iscsi_io->asc = (task->sense.ascq >> 8) & 0xFF;
 	iscsi_io->ascq = task->sense.ascq & 0xFF;
